Use the validation endpoint to get user ID
This commit causes the client to start using the validation endpoint
when setting up the client, in order to validate the access token. This
is highly encouraged by the Twitch API anyway.

Since the validation endpoint provides the user ID as part of a
successful response, then it is not necessary to manually ask the user
for the channel name and cast the user login to the user ID, because it
can be extracted by looking at who the token belongs to.

References to the configuration key for the channel login have been
removed from the config class and from the README file.