(Re)introduce -verbose flag to operate the logger
A logger has been added to the application so that further messages are
logged to the default output if the -verbose flag is added when
launching the application. If missing, the logger messages are
discarded.

Since the verbose logging already prints the location of the
configuration file, the -env flag has been stripped because it is not
necessary anymore to have a separate option to print the current config
file location.