Split application data in config and state
This commit replaces the StreamTitle.env file, previously used to store
both the client credentials and the refresh token, into a separate
config file that contains the client credentials, and a state file that
contains the current access token and refresh token of the application.

The new files use TOML, because I am planning on having more data types
in the future and I'd like to store full data structures. I'm thinking
about profiles in the config file.

For the first time, I am storing the access token to try to use it
before giving up and using the refresh token to issue a new access
token. Therefore, the validation flow when starting the application will
now try to validate the previous access token before discarding it and
issuing a new one.

After applying this commit, it will be necessary to recreate the config
file according to what the READMe file states, or the application will
not work.