Add support for profiles
As stated in the README, a profile is a prerecorded set of parameters to
update the channel information, to avoid having to always change the
same values multiple times.

This commit updates the config schema so that profiles can be stored in
the config.toml file and parsed when the application starts. Then the
profiles are available and can be picked using the new -profile flag.

The -list-profiles flag has also been introduced and it prints the list
of profiles that are present in the config file and available for use.
Using a profile that does not exist will report an error and not update
the channel information.

A -dry-run flag was introduced because it was useful during development,
and it is also probably useful during application launch.